fix: made changes raised in comments and moved radio-button types to types directory
